Track active chats and notify users of queue size
Introduces an activeChats set to monitor concurrent users and sends a notification to each user about the current number of active chats. If more than 5 users are active, an additional message warns of possible delays. Also ensures chat IDs are removed from activeChats when processing is complete, and disables the 'search' option in the prompt.